+++ src/main/java/nextstep/ladder/domain/game/Participants.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.game;
 
 import java.util.Iterator;
 import java.util.List;
