+++ src/main/java/nextstep/ladder/domain/game/exception/IllegalMaxLengthValueException.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain.exception;
+package nextstep.ladder.domain.game.exception;
 
 public class IllegalMaxLengthValueException extends IllegalArgumentException {
     public IllegalMaxLengthValueException(String message) {
