+++ src/main/java/nextstep/ladder/domain/game/exception/NoEqualLengthArgumentException.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain.exception;
+package nextstep.ladder.domain.game.exception;
 
 public class NoEqualLengthArgumentException extends IllegalArgumentException {
     public NoEqualLengthArgumentException(String message) {
