+++ src/main/java/nextstep/ladder/domain/game/exception/ParticipantNotFoundException.java
@@ -1,0 +1,7 @@
+package nextstep.ladder.domain.game.exception;
+
+public class ParticipantNotFoundException extends IllegalArgumentException {
+    public ParticipantNotFoundException(String message) {
+        super(message);
+    }
+}
