+++ src/main/java/nextstep/ladder/domain/line/LadderPointGenerator.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.line;
 
 import java.util.Random;
