+++ src/main/java/nextstep/ladder/domain/line/Point.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.line;
 
 public class Point {
     private final int index;
