+++ src/test/java/nextstep/ladder/domain/line/PointTest.java
@@ -1,5 +1,6 @@
-package nextstep.ladder.domain;
+package nextstep.ladder.domain.line;
 
+import nextstep.ladder.domain.line.Point;
 import org.junit.jupiter.api.Test;
 
 import static java.lang.Boolean.FALSE;
